requirements.txt
python-dotenv==1.0.0
langchain==0.1.16
langchain-openai==0.1.3
openai==1.18.0
langchain-groq==0.1.3
pypdf==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9